requirements.dev.txt
# main dependencies
-r requirements.txt

#dev
pytest
pytest-django
ipdb
tox
flake8
django-redis